fix wall warnings.

+++ CamStudioRecorder/application_settings_ui.cpp
@@ -143,10 +143,10 @@ void application_settings_ui::OnCbnSelchangeTempDirectoryCombobox()
 std::wstring get_window_text(const CWnd &window)
 {
     const auto window_text_length = window.GetWindowTextLengthW();
-    auto window_text = std::wstring(window_text_length + 1, '\0');
+    auto window_text = std::wstring(static_cast<size_t>(window_text_length + 1), '\0');
     // \todo make safe lowering cast..
     window.GetWindowTextW(window_text.data(), static_cast<int>(window_text.size()));
-    window_text.resize(window_text_length);
+    window_text.resize(static_cast<size_t>(window_text_length));
     return window_text;
 }
